Move Yahoo API types to dedicated types directory
Refactored type definitions for better code organization and maintainability:

• Moved all Yahoo Fantasy API interfaces from lib/yahoo-fantasy.ts to types/yahoo-fantasy.ts
• Updated imports across components and API files to use new type location
• Improved separation of concerns between type definitions and business logic
• Enhanced code reusability by allowing components to import types independently

This change follows TypeScript best practices and makes the codebase more scalable as new features are added.
