+++ src/components/players-table/player-stats-cell.tsx
@@ -1,4 +1,4 @@
-import { YahooPlayerStats } from '@/lib/yahoo-fantasy';
+import { YahooPlayerStats } from '@/types/yahoo-fantasy';
 
 interface PlayerStatsCellProps {
   player: YahooPlayerStats;
