+++ src/types/yahoo-fantasy.ts
@@ -1,0 +1,73 @@
+export interface YahooUserResponse {
+  fantasy_content: {
+    users: Array<{
+      user: Array<{
+        profile: {
+          display_name: string;
+          fantasy_profile_url: string;
+          image_url: string;
+        };
+      }>;
+    }>;
+  };
+}
+
+export interface YahooGameKey {
+  game_key: string;
+  game_id: string;
+  name: string;
+  code: string;
+  type: string;
+  season: string;
+}
+
+export interface YahooGamesResponse {
+  fantasy_content: {
+    games: Array<{
+      game: [YahooGameKey, unknown];
+    }>;
+  };
+}
+
+export interface YahooPlayerStats {
+  player_key: string;
+  name: {
+    full: string;
+    first: string;
+    last: string;
+  };
+  editorial_team_abbr: string;
+  display_position: string;
+  player_stats?: {
+    stats: Array<{
+      stat_id: number;
+      value: string | number;
+    }>;
+  };
+}
+
+export interface YahooPlayersResponse {
+  fantasy_content: {
+    game: [
+      unknown,
+      {
+        players: {
+          count: number;
+          [key: string]: {
+            player: [
+              Array<Record<string, unknown>>,
+              {
+                player_stats?: {
+                  stats: Array<{
+                    stat_id: number;
+                    value: string | number;
+                  }>;
+                };
+              }
+            ];
+          } | number;
+        };
+      }
+    ];
+  };
+}
